update : 마이페이지 UI 재배치 (#44)
