+++ src/components/mypage/MyTabs.tsx
@@ -1,29 +1,26 @@
-import React, { useState } from 'react';
-import { Tabs, Tab, Card, CardBody } from '@nextui-org/react';
-import { useQuery } from '@tanstack/react-query';
 import { getBookmarksByUserId } from '@/apis/bookmarks';
-// import { getLikesByUserId } from '@/apis/likes';
-// import ReviewCard from '../common/ReviewCard';
 import { getLikedReviews, getReviewsByUserIdrpc } from '@/apis/reviews';
 import PlaceCard2 from '../common/PlaceCard2';
 import ReviewCard2 from '../common/ReviewCard2';
+import { useQuery } from '@tanstack/react-query';
+import { Card, CardBody, Tabs, Tab } from '@nextui-org/react';
+import { useSelector } from 'react-redux';
+import { RootState } from '@/redux/config/configStore';
+import CardProfile from './CardProfile';
 
-type Props = {
-  userId: string;
-};
+const MyTabs = () => {
+  const { userId } = useSelector((state: RootState) => state.auth);
 
-const MyTabs = ({ userId }: Props) => {
   const { data: bookmarkedPlaces, isLoading: isBookmarksLoading } = useQuery({
     queryKey: ['bookmark', userId],
     queryFn: () => getBookmarksByUserId(userId),
+    enabled: !!userId,
   });
 
   const { data: likedReviews, isLoading: isLikesLoading } = useQuery({
     queryKey: ['likes', userId],
     queryFn: () => getLikedReviews(userId),
   });
-
-  console.log('좋아요 한 리뷰들>>', likedReviews);
 
   const { data: writtenReviews, isLoading: isWrittenReviewsLoading } = useQuery(
     {
@@ -41,37 +38,65 @@ const MyTabs = ({ userId }: Props) => {
 
   return (
     <div className='flex w-full flex-col'>
-      <Tabs aria-label='Options' color='primary'>
-        <Tab key='photos' title='내가 북마크한 장소'>
+      <Tabs
+        aria-label='Options'
+        color='warning'
+        className='w-full flex justify-center'
+        size='lg'
+        variant='underlined'
+      >
+        <Tab key='profile' title='프로필 수정'>
+          <CardProfile />
+        </Tab>
+        <Tab key='bookmarked' title='내가 북마크한 장소'>
           <Card>
-            <CardBody className='grid grid-cols-4 gap-12'>
+            <CardBody>
               {bookmarkedPlaces?.length !== 0 ? (
-                bookmarkedPlaces?.map((place, idx) => (
-                  <PlaceCard2 key={idx} place={place} />
-                ))
+                <div className='grid grid-cols-4 gap-12'>
+                  {bookmarkedPlaces?.map((place, idx) => (
+                    <PlaceCard2 key={idx} place={place} />
+                  ))}
+                </div>
               ) : (
-                <div>북마크한 장소가 없습니다</div>
+                <div className='flex justify-center w-full '>
+                  북마크한 장소가 없습니다 😢
+                </div>
+              )}
+            </CardBody>
+          </Card>
+        </Tab>
+        <Tab key='liked' title='내가 좋아요한 리뷰'>
+          <Card>
+            <CardBody>
+              {likedReviews?.length !== 0 ? (
+                <div className='flex flex-col gap-1'>
+                  {likedReviews?.map((review, idx) => (
+                    <ReviewCard2 review={review} key={idx} />
+                  ))}
+                </div>
+              ) : (
+                <div className='flex justify-center w-full'>
+                  좋아요한 리뷰가 없습니다. 😢
+                </div>
               )}
             </CardBody>
           </Card>
         </Tab>
 
-        <Tab key='music' title='내가 좋아요한 리뷰'>
+        <Tab key='written' title='내가 작성한 리뷰'>
           <Card>
-            <CardBody className='flex flex-col gap-1'>
-              {likedReviews?.map((review, idx) => (
-                <ReviewCard2 review={review} key={idx} />
-              ))}
-            </CardBody>
-          </Card>
-        </Tab>
-
-        <Tab key='videos' title='내가 작성한 리뷰'>
-          <Card>
-            <CardBody className='flex flex-col gap-1'>
-              {writtenReviews?.map((review, idx) => (
-                <ReviewCard2 key={idx} review={review} />
-              ))}
+            <CardBody>
+              {writtenReviews?.length !== 0 ? (
+                <div className='flex flex-col gap-1'>
+                  {writtenReviews?.map((review, idx) => (
+                    <ReviewCard2 key={idx} review={review} />
+                  ))}
+                </div>
+              ) : (
+                <div className='flex justify-center w-full'>
+                  작성한 리뷰가 없습니다. 😢
+                </div>
+              )}
             </CardBody>
           </Card>
         </Tab>
